backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
mysqlclient==2.2.0
Pillow==10.1.0
factory-boy==3.3.0
pytest==7.4.3
pytest-django==4.7.0
gunicorn==21.2.0
